requirements.txt
django==5.0.4
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.0.0
mypy==1.7.0
django-stubs[compatible-mypy]==4.2.7
coverage==7.3.0
django-bootstrap5==24.1
unittest-parametrize==1.4.0
djangosaml2==1.9.2
